feat(useTagSelection): 훅 구현 및 적용
